Made important changes in some files.
Game.java has more methods to use that are generally good for every
game.
Player.java is not abstract anymore, the reason being I changed it to a
class that stores only statistics, so each game can use it to keep the
bare minimum of stats, or expand with new ones.
I also made progress with my own game MemoryGame.java, however it's not
completed, the reason being I want to use this class has a prototype of
what the final project will look like, with OO programming,
encapsulation and polymorphism.